backend/requirements.txt
anyio==3.3.4
asgiref==3.4.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
dnspython==2.1.0
ecdsa==0.17.0
elasticsearch==7.15.1
email-validator==1.1.3
fastapi==0.70.0
google-auth==2.3.3
h11==0.12.0
idna==3.3
importlib-metadata==4.8.1
numpy==1.21.3
passlib==1.7.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pymongo==3.12.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
requests==2.26.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3
starlette==0.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
zipp==3.6.0
